+++ MatchingMakingMonitor/Models/DisplayPlayerStats.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Windows;
 using System.Windows.Media;
-using MatchMakingMonitor.View;
+using MatchMakingMonitor.View.Util;
 
 namespace MatchMakingMonitor.Models
 {
